# apps/next/public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Next REST Framework",
    "description": "This is an autogenerated documentation by Next REST Framework.",
    "version": "v4.0.0"
  },
  "paths": {
    "/api/auth/check": {
      "options": {
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsCheckResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "operationId": "",
        "tags": ["_options"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "parameters": "",
        "callbacks": "",
        "security": "",
        "servers": ""
      },
      "post": {
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCheckResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "operationId": "checkRole",
        "tags": ["auth"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "parameters": "",
        "callbacks": "",
        "security": "",
        "servers": ""
      }
    },
    "/api/auth/check/{role}": {
      "options": {
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsCheckResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "parameters": "",
        "operationId": "",
        "tags": ["_options"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "callbacks": "",
        "security": "",
        "servers": ""
      },
      "post": {
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostCheckResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostCheckErrorResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "parameters": "",
        "operationId": "hasRole",
        "tags": ["auth"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "callbacks": "",
        "security": "",
        "servers": ""
      }
    },
    "/api/auth/login": {
      "options": {
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsLoginResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "operationId": "",
        "tags": ["_options"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "parameters": "",
        "callbacks": "",
        "security": "",
        "servers": ""
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PostLoginRequestBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostLoginResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostLoginErrorResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "operationId": "login",
        "tags": ["auth"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "parameters": "",
        "callbacks": "",
        "security": "",
        "servers": ""
      }
    },
    "/api/codes": {
      "options": {
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsCodesResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "operationId": "",
        "tags": ["_options"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "parameters": "",
        "callbacks": "",
        "security": "",
        "servers": ""
      },
      "get": {
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCodesResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetCodesErrorResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "operationId": "getCodes",
        "tags": ["codes"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "parameters": "",
        "callbacks": "",
        "security": "",
        "servers": ""
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PostCodesRequestBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostCodesResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostCodesErrorResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "operationId": "createCode",
        "tags": ["codes"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "parameters": "",
        "callbacks": "",
        "security": "",
        "servers": ""
      },
      "delete": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCodesRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCodesResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCodesErrorResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "An unknown error occurred, trying again might help.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnexpectedError" }
              }
            }
          }
        },
        "operationId": "deleteCode",
        "tags": ["codes"],
        "description": "",
        "summary": "",
        "externalDocs": "",
        "deprecated": false,
        "parameters": "",
        "callbacks": "",
        "security": "",
        "servers": ""
      }
    }
  },
  "components": {
    "schemas": {
      "DeleteCodesErrorResponseBody": { "type": "string" },
      "DeleteCodesRequestBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8,
            "pattern": "^\\d+$"
          }
        },
        "required": ["code"],
        "additionalProperties": false
      },
      "DeleteCodesResponseBody": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "timesUsed": { "type": "number" },
          "value": { "type": "string" }
        },
        "required": ["id", "timesUsed", "value"],
        "additionalProperties": false
      },
      "GetCodesErrorResponseBody": { "type": "string" },
      "GetCodesResponseBody": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": { "type": "number" },
            "timesUsed": { "type": "number" },
            "value": { "type": "string" }
          },
          "required": ["id", "timesUsed", "value"],
          "additionalProperties": false
        }
      },
      "OptionsCheckResponseBody": { "type": "string" },
      "OptionsCodesResponseBody": { "type": "string" },
      "OptionsLoginResponseBody": { "type": "string" },
      "PostCheckErrorResponseBody": { "type": "string" },
      "PostCheckResponseBody": { "type": "string" },
      "PostCodesErrorResponseBody": { "type": "string" },
      "PostCodesRequestBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8,
            "pattern": "^\\d+$"
          }
        },
        "required": ["code"],
        "additionalProperties": false
      },
      "PostCodesResponseBody": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "timesUsed": { "type": "number" },
          "value": { "type": "string" }
        },
        "required": ["id", "timesUsed", "value"],
        "additionalProperties": false
      },
      "PostLoginErrorResponseBody": { "type": "string" },
      "PostLoginRequestBody": {
        "type": "object",
        "properties": { "code": { "type": "string" } },
        "required": ["code"],
        "additionalProperties": false
      },
      "PostLoginResponseBody": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "admin"] }
        },
        "required": ["token", "role"],
        "additionalProperties": false
      },
      "UnexpectedError": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "additionalProperties": false
      }
    }
  }
}